+++ .storybook/preview.tsx
@@ -6,12 +6,13 @@ import rootMessagesEn from '../src/rootMessages.en';
 import { FONT_CUSTOMIZATIONS } from '../src/constants/fontCustomizations';
 import { Global } from '@emotion/react';
 import { GLOBAL_STYLE } from '../src/constants/globalStyle';
+import { FONT_STYLE } from '../src/constants/fontStyle';
 
 const preview: Preview = {
   decorators: [
     (Story) => (
       <>
-        <Global styles={{ GLOBAL_STYLE }} />
+        <Global styles={{ GLOBAL_STYLE, FONT_STYLE }} />
         <IntlProvider messages={rootMessagesEn} locale="en">
           <BoemlyThemeProvider fonts={FONT_CUSTOMIZATIONS}>
             <Story />
